pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghl-base"
version = "0.1.0"
description = "GoHighLevel base: OAuth flow and aiohttp webhook server"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "aiohttp>=3.9",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]
authors = [{name = "GHL Base", email = "dev@example.com"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.scripts]
ghl-webhooks = "ghl_base.webhook_app:main"
ghl-oauth = "ghl_base.oauth:main"

[tool.setuptools.packages.find]
where = ["src"]
